[GEOS-9311] Switching from System Status to Modules tab gives an error
