+++ src/web/core/src/main/java/org/geoserver/web/system/status/SystemStatusMonitorPanel.html
@@ -2,24 +2,7 @@
 <head></head>
 <body>
 	<wicket:panel>
-		<div>
-			<wicket:message key="updated">Updated at</wicket:message>
-			<span wicket:id="time"></span>
-		</div>
-		<table>
-			<thead>
-				<tr>
-					<th><span><wicket:message key="info">Info</wicket:message></span></th>
-					<th><span><wicket:message key="value">Value</wicket:message></span></th>
-				</tr>
-			</thead>
-			<tbody>
-				<tr wicket:id="metrics">
-					<th scope="row"><span wicket:id="info"></span></th>
-					<td><span wicket:id="value"></span></td>
-				</tr>
-			</tbody>
-		</table>
+		<div wicket:id="refreshed-values"></div>
 	</wicket:panel>
 </body>
 </html>
